requirements.txt
# Web Framework
Flask==3.1.2
flask-cors==6.0.1

# AI/ML Core Libraries
openai==2.6.1
sentence-transformers==5.1.2

# Vector Database
faiss-cpu==1.12.0

# LangChain Ecosystem (for langchain_test.py helper script)
langchain-huggingface==1.0.0
langchain-community==0.4.1

# Data Processing
pandas
numpy==2.3.4

# Turkish NLP
zemberek-grpc==0.16.1
snowballstemmer==3.0.1

# gRPC
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf>=5.26.1,<6.0

# Utilities
python-dotenv==1.2.1
PyYAML==6.0.2